Avoid the need to allocate memory for addresses
In the old code, the fill_segment function always dynamically allocated memory for the pseudo-random addresses used in Argon2i. However, the addresses can be easily computed on-the-fly, thus avoiding the need for dynamic allocation.

Another serious issue with the old code is that if the allocation failed, the fill_segment function would just silently return and do nothing. (WTF?!)

This commit refactors the code to compute the addresses on-the-fly instead of pre-computing them and also slightly simplifies the way they are computed.