Fix Travis failures
Argon2 by default runs Travis CI in a containerized environment
https://docs.travis-ci.com/user/reference/overview/#Container-based

This is faster and works fine for most things. However, #89 added ASAN
to the test cases. As ASAN requires the SYS_PTRACE capability, when
Travis/Docker permissions changed earlier this year, our tests stoped
working. See https://github.com/travis-ci/travis-ci/issues/9033

The solution is just to run ASAN in a VM (sudo=true). This test takes so
long normally that the cost to startup a VM is relativly small.