Validate lanes before memory cost.
The validation of memory cost uses a multiple of the amount of lanes.
It would be an unsigned overflow, therefore not critical, especially
since the lane validation would bail out anyway.

But an invalid lane amount would lead to the error return value
ARGON2_MEMORY_TOO_LITTLE which is not exactly what went wrong.

Therefore bail out with more specific error return value.