Fix OOB write with crazy command line argument.
This only affects:
- argon2 binary
- with crazy command line argument
- on 64 bit systems
- with plenty of RAM

How to reproduce (with a lot of memory and time):
$ echo poc | argon2 12345678 -l 4294967295

The problem is within data type handling in C.

The variable "outlen" is a uint32_t variable. The command line argument
"-l" is converted with strtoul. This by itself means that values larger
than 2**32-1 are truncated on 64 bit systems with an unsigned long being
64 bit, but let's skip this for now.

Later on in run() we see this line:
    out = malloc(outlen + 1);

The variable outlen is a function argument of type uint32_t. If the
variable contains the value 2**32-1, i.e. UINT_MAX, the addition
overflows and leads to 0. Even though malloc takes a size_t, the C
language only performs this integer calculation with 32 bit, unsigned.

Yes, requesting a hash output length of 4 GB is pretty stupid and
your system takes more than that, since base64 encoding and internally
used memory by argon2 library functions will likely lead to an error
sooner or later.

If the code eventually completes, argon2_hash would try to memcpy the
raw hash into the 0-byte allocated memory area. Since the raw hash is
4 GB and not 0, an OOB write and a basically instant crash occur.

Defining outlen as size_t AND validating that the calculation won't
overflow by bailing out early on invalid command line argument fixes the
problem.